เสร็จสิ้นการคืนรางวัลให้ผู้เล่นได้และเกมส์สามารถคำนวณหาผู้ชนะได้

+++ src/CasinovaAll/Modules/MagicNine/TheS.Casinova.MagicNine.DAL.BackContracts/Commands/GetGameRoundPotCommand.cs
@@ -23,6 +23,6 @@ namespace TheS.Casinova.MagicNine.Commands
         /// <summary>
         /// เงินในกอง ที่ต้องการ
         /// </summary>
-        public int Pot { get; set; }
+        public double Pot { get; set; }
     }
 }
